requirements_test.txt
pytest>=7.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0
pytest-homeassistant-custom-component>=0.13.296
aiohttp>=3.8.0,<4.0.0
mypy>=1.0.0
homeassistant-stubs>=2025.12.0
ruff>=0.1.0
pre-commit>=3.0.0
